requirements.txt
langchain
langchain-community
langchain-huggingface
groq
chromadb
pandas
streamlit
python-docx
python-dotenv